pyproject.toml
[tool.poetry]
name = "junctionbp2023"
version = "0.1.0"
description = ""
authors = ["Marcell O <omarcell@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.3.5"
logging = "^0.4.9.6"
python-dotenv = "^1.0.0"
llama-cpp-python = "^0.2.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
